requirements.txt
streamlit>=1.32
python-dotenv>=1.0
openai>=1.0
pydantic>=2.0
tzdata>=2024.1
